Remove kitchen spike component references
