--- Content.Client/Kitchen/Components/KitchenSpikeComponent.cs
@@ -1,9 +1,0 @@
-using Content.Shared.Kitchen.Components;
-
-namespace Content.Client.Kitchen.Components
-{
-    [RegisterComponent, ComponentReference(typeof(SharedKitchenSpikeComponent))]
-    public sealed class KitchenSpikeComponent : SharedKitchenSpikeComponent
-    {
-    }
-}
